Revert "Merge branch 'developer' into CustomerDetails"
This reverts commit f9ed5dcd91ffdbb6362fc9b5b3b64a819903e909, reversing
changes made to 73cb1f7b4a07c5c47ae7990f056a90005d8fb1fa.

+++ BackEnd/src/main/java/septmajorproject/bookingsys/repository/CustomerRepository.java
@@ -1,15 +1,16 @@
 package septmajorproject.bookingsys.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import septmajorproject.bookingsys.model.Customer;
 
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer,Integer> {
     @Override
-        //returns a list of all customers
+    //returns a list of all customers
     List<Customer> findAll();
 
     //finds customer by unique username
